Organize components into folders based on where they are used
For example, dashboard components go into the dashboard folder,
settings components go into the settings folder, and components
used in multiple places go into the common folder.
