+++ src/components/common/Footer.test.js
