+++ src/components/common/StatusBadge.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import { translate } from 'react-i18next';
-import { api, makeCancelable, ignoreCancel } from '../utils';
+import { api, makeCancelable, ignoreCancel } from '../../utils';
 
 class StatusBadge extends Component {
   state = {
