+++ src/components/dashboard/ClientsGraph.js
@@ -12,7 +12,7 @@ import React, { Component } from 'react';
 import ReactDOM from 'react-dom';
 import { Line } from 'react-chartjs-2';
 import { translate } from 'react-i18next';
-import { padNumber, api, makeCancelable, ignoreCancel } from '../utils';
+import { padNumber, api, makeCancelable, ignoreCancel } from '../../utils';
 import ChartTooltip from "./ChartTooltip";
 
 class ClientsGraph extends Component {
