+++ src/components/list/DomainInput.test.js
@@ -11,7 +11,7 @@
 import React from 'react';
 import { shallow } from 'enzyme';
 import DomainInput from './DomainInput';
-import { api } from "../utils";
+import { api } from "../../utils";
 
 it('has a placeholder', () => {
   const placeholder = 'placeholder';
