+++ src/components/settings/VersionCard.js
